Combined both lines of work without dropping either:
- kept our backend: graceful shutdown, resource sampler, reconnect, idempotency, Disruptor
- kept colleague's: console-api service, console-ui/leaderboard-ui front-ends, env-config, per-sandbox network isolation
- unified leaderboard Entry (union of fields), docker sandbox (sampler + network), docs

+++ services/leaderboard-api/internal/redisboard/redisboard.go
@@ -328,9 +328,9 @@ func atof(s string, def float64) float64 {
 	return v
 }
 
-func atoi(s string) int64 {
+func atoi(s string) int {
 	v, _ := strconv.ParseInt(s, 10, 64)
-	return v
+	return int(v)
 }
 
 func boolToInt(b bool) int {
